Implemented IdP selection
The user is now able to select an IdP when directed to the discovery
service. This is implemented as a form on the groups page. Once the form
is submitted, a redirect is made back to the SP with the selected
EntityID (via path variable). The SP should then redirect the
user to the IdP for authentication.