Reworked entity storage plus refactor
Now entities are stored in redis as a hash keyed by entity ID. This
occurred because the Sinatra application requires fast access to SP data
when serving up a request. There was the option of storing each record
individually in redis (rather than using an hash per group), however this
proved to be complex to track changes — this is future a requirement for
DS.

Also, several refactoring efforts were made during this process.