Added support for passive parameter
If the service provider included the isPassive parameter in its
original request, then the discovery service has no option and MUST redirect
the user agent back to the service provider.

If it responds, then it MUST NOT include the query string parameter
whose name was specified as the value of the returnIDParam parameter in
the original request (or entityID if no parameter was supplied). The
absence of this parameter is the indication of failure to return a selection.

 If isPassive is true, the IdP may rely on saved state such as cookies
 to redirect to the IdP