Introduce sprockets with JQuery / Semantic UI
This allows us to precompile all web assets in a single (css and js) file to
improve pageload latency.

Multiple instance of the sprockets pipeline will exist in the Sinatra app AND
update process. This is so the digest values for the linked stylesheet /
js stay the same (the update process creates a page to be used later).