Implemented basic datatable
After some experimentation, a JS data driven approach is easier to
manage.

Rendering multiple tabs in Slim was messy as several instances of
datatables were required. The page's model was mixed throughout the HTML
and JS. This was complex to work with. Geolocation, entity IDs and
logos were already in the page (via JSON array) anyway. It made sense to use
this as the data source.