Reworked passive logic to GET /discovery/:group
Fixes https://github.com/ausaccessfed/discovery-service/issues/25.

Follows isPassive rules outlined in this document:
http://docs.oasis-open.org/security/saml/Post2.0/sstc-saml-idp-discovery.pdf

Small note: there is an edge case:
"If the service provider included the isPassive parameter in its original
request, then the discovery service has no option and MUST redirect the user
agent back to the service provider. If it responds, then it MUST NOT include
the query string parameter whose name was specified as the value of the
returnIDParam parameter in the original request (or entityID if no parameter
was supplied). The absence of this parameter is the indication of failure to
return a selection."

I've catered for this however I'm not recording it in the audit trail.
The user isn't making an IdP selection so I'm unsure what value this will
add. I'm open to discussion though.