Implemented keyboard support for non JS users
This implementation is basic yet functional. It allows non JS users to
select an IdP without a mouse (using tab key to select IdP, then enter).

There is no support for "Remember your IdP" as you need to press enter to
select this field (this submits the form). Every solution I've encountered
uses Javascript to avoid this.

The other option would be to use <noscript> tags but this might be
be outside the scope of what we're trying to achieve.