Correct text of strings that are HTML encoded
Data which is provided to JSON parse from `#idps` or `#sps` is
still HTML encoded. For example ' is represented by `&#39;`. Whilst this
is safe enough in terms of valid JSON strings it is errenous when shown
to end users e.g. "TRS&#39;s SP" is far from ideal.

This change ensures that strings which may contain this extra data is
corrected decoded. Furthermore use of `textArea` reduces the chance of
XSS injection via metadata source.

I did attempt to apply the decode process to the entire string prior to
JSON parsing but this unfortunately breaks in the brower for reasons I
was unable to fully correct, hence this approach.